+++ lib/facebook_app_events.dart
@@ -11,12 +11,19 @@ class FacebookAppEvents {
   // See: https://github.com/facebook/facebook-android-sdk/blob/master/facebook-core/src/main/java/com/facebook/appevents/AppEventsConstants.java
   static const eventNameActivatedApp = 'fb_mobile_activate_app';
   static const eventNameDeactivatedApp = 'fb_mobile_deactivate_app';
-  static const eventNameCompletedRegistration = 'fb_mobile_complete_registration';
+  static const eventNameCompletedRegistration =
+      'fb_mobile_complete_registration';
   static const eventNameViewedContent = 'fb_mobile_content_view';
   static const eventNameRated = 'fb_mobile_rate';
+  static const eventNameInitiatedCheckout = 'fb_mobile_initiated_checkout';
 
   static const _paramNameValueToSum = "_valueToSum";
+  static const paramNameCurrency = "fb_currency";
   static const paramNameRegistrationMethod = "fb_registration_method";
+  static const paramNamePaymentInfoAvailable = "fb_payment_info_available";
+  static const paramNameNumItems = "fb_num_items";
+  static const paramValueYes = "1";
+  static const paramValueNo = "0";
 
   /// Parameter key used to specify a generic content type/family for the logged event, e.g.
   /// "music", "photo", "video".  Options to use will vary depending on the nature of the app.
@@ -50,6 +57,10 @@ class FacebookAppEvents {
   /// Returns the app ID this logger was configured to log to.
   Future<String> getApplicationId() {
     return _channel.invokeMethod<String>('getApplicationId');
+  }
+
+  Future<String> getAnonymousId() {
+    return _channel.invokeMethod<String>('getAnonymousId');
   }
 
   /// Log an app event with the specified [name] and the supplied [parameters] value.
@@ -190,18 +201,6 @@ class FacebookAppEvents {
     );
   }
 
-  /// Creates a new map containing all of the key/value pairs from [parameters]
-  /// except those whose value is `null`.
-  Map<String, dynamic> _filterOutNulls(Map<String, dynamic> parameters) {
-    final Map<String, dynamic> filtered = <String, dynamic>{};
-    parameters.forEach((String key, dynamic value) {
-      if (value != null) {
-        filtered[key] = value;
-      }
-    });
-    return filtered;
-  }
-
   /// Re-enables auto logging of app events after user consent
   /// if disabled for GDPR-compliance.
   ///
@@ -236,7 +235,11 @@ class FacebookAppEvents {
   /// This is needed for California Consumer Privacy Act (CCPA) compliance
   ///
   /// See: https://developers.facebook.com/docs/marketing-apis/data-processing-options
-  Future<void> setDataProcessingOptions(List<String> options, { int country, int state }) {
+  Future<void> setDataProcessingOptions(
+    List<String> options, {
+    int country,
+    int state,
+  }) {
     final args = <String, dynamic>{
       'options': options,
       'country': country,
@@ -245,13 +248,56 @@ class FacebookAppEvents {
 
     return _channel.invokeMethod<void>('setDataProcessingOptions', args);
   }
-  
-  Future<void> logPurchase({@required double amount, @required String currency, Map<String, dynamic> parameters}) {
+
+  Future<void> logPurchase({
+    @required double amount,
+    @required String currency,
+    Map<String, dynamic> parameters,
+  }) {
     final args = <String, dynamic>{
       'amount': amount,
       'currency': currency,
       'parameters': parameters,
     };
     return _channel.invokeMethod<void>('logPurchase', _filterOutNulls(args));
+  }
+
+  Future<void> logInitiatedCheckout({
+    @required double totalPrice,
+    @required String currency,
+    @required String contentType,
+    @required String contentId,
+    @required int numItems,
+    bool paymentInfoAvailable = false,
+  }) {
+    return logEvent(
+      name: eventNameInitiatedCheckout,
+      valueToSum: totalPrice,
+      parameters: {
+        paramNameContentType: contentType,
+        paramNameContentId: contentId,
+        paramNameNumItems: numItems,
+        paramNameCurrency: currency,
+        paramNamePaymentInfoAvailable:
+            paymentInfoAvailable ? paramValueYes : paramValueNo,
+      },
+    );
+  }
+
+  // ---------------------------------------------------------------------------
+  // ---------------------------------------------------------------------------
+  //
+  // PRIVATE METHODS BELOW HERE
+
+  /// Creates a new map containing all of the key/value pairs from [parameters]
+  /// except those whose value is `null`.
+  Map<String, dynamic> _filterOutNulls(Map<String, dynamic> parameters) {
+    final Map<String, dynamic> filtered = <String, dynamic>{};
+    parameters.forEach((String key, dynamic value) {
+      if (value != null) {
+        filtered[key] = value;
+      }
+    });
+    return filtered;
   }
 }